fix renaming transactionform
